A long message to explain a major re-thinking of the various lexer classification tokens.
Dividing the world of names into ident (purely alphanumerical) and symbols
(containing some non alphanumerical characters).

Abstracting creation of tokens (for a more consistent usage and naming).
Renaming some identifiers to more significant ones.
Documentation. Typos.

First trial to understand the implications for the lexer of the necessary
identifier classifications for the parser. For instance, how the lexer must
deal with stratification between module and regular value names ? Or
constructor versus regular value names ? How can the lexer (properly) deal
with ``fixity'' status of names (infix/prefix/always prefix) ? How can the
lexer deal with the big world of names division (alphanumerical/symbolic) ?
Last but not least, how to deal with the cartesian product of all those
(necessary) categorization ``plans'' ?

To fix the vocabulary, let's name:
 - ``Symbolism'' the big name partition between alphanumeric and symbolic names,
 - ``Fixity'' the syntactic partition between infix/prefix/always prefix of names,
 - ``Syntaxity'' the syntactic partition between value/constructor, type/module ,
    collection/value, and function/species syntactic category of names.

Then we have to define how the lexer must divide tokens into those various
partitions, i.e. give them a value in the cartesian product

 Symbolism x Fixity x Syntaxity.

In the first place, we need to define those various partitions to get a true
partition (in the mathematical sense) of the set of names.

For the time being in Focalize, the partitions are more or less defined as:
 - Symbolism is ``purely alphanumerical''/``symbolic''.
 - Fixity is ``infix''/``always prefix''/``regular''.
 - Syntaxity is ``lowercase''/``uppercase''.

Simple examples:
 - ``x'' is (``purely alphanumerical'', ``regular'', ``lowercase'').
 - ``True'' is (``purely alphanumerical'', ``regular'', ``uppercase'').

 - ``+'' is ``symbolic'' and ``infix''; however, its syntaxity is not
   explicit in the lexer/parser (may be we can treat this as an extra ``ad
   hoc'' class of syntaxity ? Mmm, ugly!).

Trickier:
 - ``::'' is (``symbolic'', ``infix'', ``uppercase'').
   In effect:
   * ``::'' is not alphanumerical, hence is ``symbolic'';
   * ``::'' is a constructor, hence is ``uppercase'';
   * ``::'' is an infix, hence is ``infix''.

 - ``()'' and ``[]'' are (``symbolic'', ``regular'', ``uppercase'').

 - ``+'' is (``symbolic'', ``infix'', ``lowercase'').
   In effect:
   * ``+'' is not ``uppercase'' (it cannot be used as a constructor name),
     hence it  lowercase. For the same reason:
 - ``~'' is (``symbolic'', ``always prefix'', ``lowercase'').

Relationship between those lexical partitions is not explicit if always
clear; furthermore, it has some ``rugosity'' to be worked out.