Fix bug 18 and invisible bug in rule PRM. Now, in PRM, we look for "z's" in BODY but *also* in TYPE. This was bug 18.
Moreover, we fixed the lookup of "y's" by doing it only
in DIDOU ([TYPE] + [PRM]) (the same set that the one used to compute things
to lambda-lift for the record type) instead of in all the dependencies.
About this last point, I have a strong conviction that at least PRM is not
needed, and also not needed (by the way) when computing what to lambda-lift for
the record type.