[Focalide] Print arrows as (cc.eT A -> cc.eT B) instead of cc.eT (cc.Arrow A B)
This is more readable and allow for less printing of dk_builtins.prop,
the code of Prop in type cc.uT.

If we do not use prop for typing arguments (especially if we restrict to
first-order logic), dk_builtins.prop is not used and can be removed from
the file dk_builtins.prop.  It is for now still there because the
standard library contains a second-order induction principle.

Removing dk_builtins.prop allows to interpret prop in a universe
different from cc.uT, as is the case when interoperating with Coq.

If we also removed cc.Arrow, we would restrict to first-order logic but
this seems too much.