Allow Zenon and Zenon Modulo to be out of the PATH
This is to allow very recent versions of Dedukti with very old
versions of Coq. Due to a current bug, the current dev version of
Dedukti requires OCaml >= 4.03 while Coq v8.4pl6 and before are not
available for this version of OCaml (at least on Opam).

FoCaLiZe allowed configuration of the PATH to Zenon but this did not
really work.